Fixed sign-in/out states not clearing
When user logs in and logs out, the email and password of the last
logged in user is displayed on the login screen. Also, after logging
into another account, the details of the previous user was not purged.
This commit fixes these problems by resetting the navigation stack
states after successful login/out.